requirements.txt
amqp==1.4.9
antiorm==1.2.1
anyjson==0.3.3
asgiref==3.3.0
beautifulsoup4==4.9.3
billiard==3.3.0.23
boto3==1.16.10
botocore==1.19.10
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
ConfigArgParse==1.4
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
Django==3.0.3
django-axes==5.8.0
django-celery==3.3.1
django-crispy-forms==1.9.2
django-elasticsearch-dsl==7.2.0
django-extensions==3.1.2
django-ipware==3.0.1
django-session-timeout==0.1.0
django-storages==1.10.1
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
fpdf==1.7.2
gensim==3.8.3
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
gunicorn==20.0.4
htmldocx==0.0.4
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kiwisolver==1.3.1
kombu==3.0.37
locust==1.4.4
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
msgpack==1.0.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psutil==5.8.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.4
pyzmq==22.0.3
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.3
sendgrid==6.4.7
six==1.15.0
smart-open==3.0.0
soupsieve==2.2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wordcloud==1.8.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
